# services/@orbital/admin-gateway/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/ping": {
      "get": {
        "operationId": "PingController_ping",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ping"
        ]
      }
    },
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{username}": {
      "get": {
        "operationId": "UsersController_findByUsername",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/areas": {
      "get": {
        "operationId": "AreasController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "areas"
        ]
      },
      "post": {
        "operationId": "AreasController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "areas"
        ]
      }
    },
    "/areas/{id}": {
      "get": {
        "operationId": "AreasController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "areas"
        ]
      },
      "put": {
        "operationId": "AreasController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "areas"
        ]
      },
      "delete": {
        "operationId": "AreasController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "areas"
        ]
      }
    },
    "/areas/{id}/map": {
      "get": {
        "operationId": "AreasController_getMap",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "areas"
        ]
      }
    }
  },
  "info": {
    "title": "Admin Gateway API",
    "description": "The Admin Gateway API provides access to administrative functions",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "admin",
      "description": ""
    },
    {
      "name": "areas",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {}
  }
}